fix(rust, python): merge rev-maps when building list arrays of categoricals. (#9742)
