+++ polars/polars-lazy/Cargo.toml
@@ -107,7 +107,7 @@ list_eval = []
 cumulative_eval = []
 chunked_ids = ["polars-plan/chunked_ids", "polars-core/chunked_ids"]
 list_to_struct = ["polars-plan/list_to_struct"]
-python = ["pyo3", "polars-plan/python"]
+python = ["pyo3", "polars-plan/python", "polars-core/python"]
 row_hash = ["polars-plan/row_hash"]
 string_justify = ["polars-plan/string_justify"]
 string_from_radix = ["polars-plan/string_from_radix"]
